Simplify Dockerfile
A previous change to the Dockerfile used in CI added a number of dependencies.
It did this both in a way that built from source instead of installing
packages, installed packages but never cleaned up the cache leaving a bloated
image, and ran a separate command leading to an extra layer in the docker
image.  This change simplifies all of those down into the three line change
that was required.