Update Dockerfile
Previously the docker file used to build the jvmkill binary was thought to be
based on a trusty image at its root. Recently it was discovered that it was
actually based on Debian Jessie which is a bit later.  This resulted in
problems around which compiler was used during the binary compilation phase.
This change updates the root of the Docker image to point at the
`cloudfoundry/cflinuxfs2` image which should match the container that the
binary will be run in. This should also ensure that the proper compiler and
ABI are used during compilation.