Reduce expectations for parallel test
The test expected to find the following output:

ResourceExhausted! (1/2)

but the output produced was:

ResourceExhausted! (OpenJDK 64-Bit Server VM warning: 1/2)

Since the JVM is concurrently writing to standard error, this test is on a
hiding to nothing. Reducing the number of expectations to increase the
probability of the test passing.