More Intelligently Find Test Agent
This change updates the code that finds the test agent during test time.
Previously, it was assumed that the first path on the library path would
contain the agent, but as of recently this is no longer true.  Now, we iterate
over every directory on the library path looking for the library.  The first
one that exists will be returned.

[#15]