Tighten Glob
Previously, only a single library file was created with the jvmkill name
during deployment.  We used a glob to cover both the macOS and linux case
(.dylib and .so) during upload, but recent changes to Rust/Cargo mean that
there's now a .d (dependency list) file of the same name.  This change
tightens the glob file to ensure that only the library is ever used.

[cloudfoundry/java-buildpack-dependency-builder#37]