Modernize
This change is a massive one.  It modernizes the code base pretty
significantly, optimizing for the latest version of Rust as well as full
testability.  I'm not gonna lie, this was really really hard and I've been
working on it part time for three months.  Let's hope this leads to more
stability in the tests and in production.

Signed-off-by: Ben Hale <bhale@pivotal.io>