Modifies the thread exhaustion test to fail more reliably
Previously, the thread exhaustion test would consume as many threads as possible until the system refused to create any more. This could sometimes crash or make the system in which the test is running unresponsive. The changes in this commit use JNA to call `setrlimit` to set the NPROC value allowed by the process. On Linux systems where threads are treated as processes this can cause the system to reject creating more threads sooner, which makes the test easier to trigger and less flaky. This won't work on non-Linux systems, so the test has a max threads of 1000 at which point the test will exit and fail.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>